refractor: move related component to grid
